+++ sdk/cosmos/azure-cosmos-spark_3_2-12/test-databricks/databricks-notebooks-install.sh
@@ -2,17 +2,19 @@
 
 CLUSTER_NAME=$1
 NOTEBOOKSFOLDER=$2
-COSMOSENDPOINT=$3
-COSMOSKEY=$4
-SUBSCRIPTIONID=$5
-TENANTID=$6
-RESOURCEGROUPNAME=$7
-CLIENTID=$8
-CLIENTSECRET=$9
-COSMOSCONTAINERNAME=${10}
-COSMOSDATABASENAME=${11}
+COSMOSENDPOINTMSI=$3
+COSMOSENDPOINT=$4
+COSMOSKEY=$5
+SUBSCRIPTIONID=$6
+TENANTID=$7
+RESOURCEGROUPNAME=$8
+CLIENTID=$9
+CLIENTSECRET=${10}
+COSMOSCONTAINERNAME=${11}
+COSMOSDATABASENAME=${12}
 [[ -z "$CLUSTER_NAME" ]] && exit 1
 [[ -z "$NOTEBOOKSFOLDER" ]] && exit 1
+[[ -z "$COSMOSENDPOINTMSI" ]] && exit 1
 [[ -z "$COSMOSENDPOINT" ]] && exit 1
 [[ -z "$COSMOSKEY" ]] && exit 1
 [[ -z "$SUBSCRIPTIONID" ]] && exit 1
@@ -23,19 +25,25 @@ COSMOSDATABASENAME=${11}
 [[ -z "$COSMOSCONTAINERNAME" ]] && exit 1
 [[ -z "$COSMOSDATABASENAME" ]] && exit 1
 
-CLUSTER_ID=$(databricks clusters list --output json | jq -r --arg N "$CLUSTER_NAME" '.clusters[] | select(.cluster_name == $N) | .cluster_id')
+CLUSTER_ID=$(databricks clusters list --output json | jq -r --arg N "$CLUSTER_NAME" '.[] | select(.cluster_name == $N) | .cluster_id')
 
 echo "Deleting existing notebooks"
-databricks workspace rm --recursive /notebooks
+if databricks workspace list / | grep -q notebooks; then
+  databricks workspace delete --recursive /notebooks
+fi
 
 echo "Importing notebooks from $NOTEBOOKSFOLDER"
-databricks workspace import_dir $NOTEBOOKSFOLDER /notebooks
+databricks workspace import-dir "$NOTEBOOKSFOLDER" /notebooks
 
-NOTEBOOKS=$(databricks workspace ls /notebooks)
+echo "Validating Notebooks in workspace"
+databricks workspace list /notebooks -o json
+
+NOTEBOOKS=$(databricks workspace list /notebooks -o json | jq -r '.[].path')
 for f in $NOTEBOOKS
 do
-	echo "Creating job for $f"
-	JOB_ID=$(databricks jobs create --json "{\"existing_cluster_id\": \"$CLUSTER_ID\", \"name\": \"$f\",\"notebook_task\": { \"notebook_path\": \"/notebooks/$f\" }}" | jq -r '.job_id')
+	name="${f##*/}"
+	echo "Creating job for $f - $name"
+	JOB_ID=$(databricks jobs create --json "{\"name\": \"$name\", \"tasks\": [{\"task_key\": \"${name}_task\", \"existing_cluster_id\": \"$CLUSTER_ID\", \"notebook_task\": { \"notebook_path\": \"$f\" }}]}" | jq -r '.job_id')
 
 	if [[ -z "$JOB_ID" ]]
 	then
@@ -44,21 +52,22 @@ do
 	fi
 
 	echo "Creating run for job $JOB_ID"
-	RUN_ID=$(databricks jobs run-now --job-id $JOB_ID --notebook-params "{\"cosmosEndpoint\": \"$COSMOSENDPOINT\",\"cosmosMasterKey\": \"$COSMOSKEY\",\"subscriptionId\": \"$SUBSCRIPTIONID\",\"tenantId\": \"$TENANTID\",\"resourceGroupName\": \"$RESOURCEGROUPNAME\",\"clientId\": \"$CLIENTID\",\"clientSecret\": \"$CLIENTSECRET\",\"cosmosContainerName\": \"$COSMOSCONTAINERNAME\", \"cosmosDatabaseName\": \"$COSMOSDATABASENAME\"}" | jq -r '.run_id')
+	echo "DBG_JSON: {\"job_id\": \"$JOB_ID\", \"notebook_params\": {\"cosmosEndpointMsi\": \"$COSMOSENDPOINTMSI\",\"cosmosEndpoint\": \"$COSMOSENDPOINT\",\"cosmosMasterKey\": \"$COSMOSKEY\",\"subscriptionId\": \"$SUBSCRIPTIONID\",\"tenantId\": \"$TENANTID\",\"resourceGroupName\": \"$RESOURCEGROUPNAME\",\"clientId\": \"$CLIENTID\",\"clientSecret\": \"$CLIENTSECRET\",\"cosmosContainerName\": \"$COSMOSCONTAINERNAME\", \"cosmosDatabaseName\": \"$COSMOSDATABASENAME\"}}"
+	RUN_ID=$(databricks jobs run-now --json "{\"job_id\": \"$JOB_ID\", \"notebook_params\": {\"cosmosEndpointMsi\": \"$COSMOSENDPOINTMSI\",\"cosmosEndpoint\": \"$COSMOSENDPOINT\",\"cosmosMasterKey\": \"$COSMOSKEY\",\"subscriptionId\": \"$SUBSCRIPTIONID\",\"tenantId\": \"$TENANTID\",\"resourceGroupName\": \"$RESOURCEGROUPNAME\",\"clientId\": \"$CLIENTID\",\"clientSecret\": \"$CLIENTSECRET\",\"cosmosContainerName\": \"$COSMOSCONTAINERNAME\", \"cosmosDatabaseName\": \"$COSMOSDATABASENAME\"}}" | jq -r '.run_id')
 
 	if [[ -z "$RUN_ID" ]]
 	then
-		echo "Could not run job"
-		databricks jobs delete --job-id $JOB_ID
+		echo "Could not run job $JOB_ID"
+		databricks jobs delete $JOB_ID
 		exit 1
 	fi
 
-	JOB_STATE=$(databricks runs get --run-id $RUN_ID | jq -r '.state.life_cycle_state')
+	JOB_STATE=$(databricks jobs get-run $RUN_ID | jq -r '.state.life_cycle_state')
 
 	if [[ -z "$JOB_STATE" ]]
 	then
 		echo "Could not find state for job $JOB_ID"
-		databricks jobs delete --job-id $JOB_ID
+		databricks jobs delete $JOB_ID
 		exit 1
 	fi
 
@@ -66,20 +75,20 @@ do
 	while [[ "$JOB_STATE" != "INTERNAL_ERROR" && "$JOB_STATE" != "TERMINATED" ]]
 	do
 		echo "Run $RUN_ID is on state '$JOB_STATE'"
-		JOB_STATE=$(databricks runs get --run-id $RUN_ID | jq -r '.state.life_cycle_state')
+		JOB_STATE=$(databricks jobs get-run $RUN_ID | jq -r '.state.life_cycle_state')
 		sleep 10
 	done
 
-	JOB_MESSAGE=$(databricks runs get --run-id $RUN_ID | jq -r '.state.state_message')
+	JOB_MESSAGE=$(databricks jobs get-run $RUN_ID | jq -r '.state.state_message')
 
 	if [[ "$JOB_STATE" != "TERMINATED" ]]
 	then
 		echo "Run $RUN_ID failed with state $JOB_STATE and $JOB_MESSAGE"
-		databricks jobs delete --job-id $JOB_ID
+		databricks jobs delete $JOB_ID
 		exit 1
 	fi
 
-	JOB_RESULT=$(databricks runs get --run-id $RUN_ID | jq -r '.state.result_state')
+	JOB_RESULT=$(databricks jobs get-run $RUN_ID | jq -r '.state.result_state')
 
 	echo "Run $RUN_ID finished with state $JOB_STATE and result $JOB_RESULT with message $JOB_MESSAGE"
 	if [[ "$JOB_RESULT" != "SUCCESS" ]]
@@ -88,5 +97,5 @@ do
 		exit 1
 	fi
 
-	databricks jobs delete --job-id $JOB_ID
+	databricks jobs delete $JOB_ID
 done