+++ sdk/elastic/azure-resourcemanager-elastic/src/main/java/com/azure/resourcemanager/elastic/models/DeploymentInfoes.java
@@ -12,7 +12,7 @@ import com.azure.core.util.Context;
  */
 public interface DeploymentInfoes {
     /**
-     * Fetch information regarding Elastic cloud deployment corresponding to the Elastic monitor resource.
+     * Fetch detailed information about Elastic cloud deployments corresponding to the Elastic monitor resource.
      * 
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param monitorName Monitor resource name.
@@ -26,7 +26,7 @@ public interface DeploymentInfoes {
     Response<DeploymentInfoResponse> listWithResponse(String resourceGroupName, String monitorName, Context context);
 
     /**
-     * Fetch information regarding Elastic cloud deployment corresponding to the Elastic monitor resource.
+     * Fetch detailed information about Elastic cloud deployments corresponding to the Elastic monitor resource.
      * 
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param monitorName Monitor resource name.